(Broken) Added Swift temp urls support
The code should be correct for generating tempurls from the given direct_urls
by glance. Also, added the necessary configs for testing. However the
generated URLs give:
401 Unauthorized: Temp URL invalid

Also, the tests for list_images() are broken because the URLs are hashed
with the time for the temp url, so they change. I'm not sure how to fix it.